bees: scene serialization fixes (include all names and versions)

+++ apps/bees/src/app_bees.h
@@ -5,7 +5,7 @@
 typedef struct _AppVersion { 
   u8 maj;
   u8 min; 
-  u8 rev;
+  u16 rev;
 } AppVersion;
 
 const AppVersion  beesVersion;
